refactoring

+++ src/main/java/com/example/my_books_backend/repository/book/BookRepository.java
@@ -11,24 +11,23 @@ import com.example.my_books_backend.entity.Book;
 
 @Repository
 public interface BookRepository extends JpaRepository<Book, String>, BookRepositoryCustom {
-       // 書籍一覧取得
-       Page<Book> findByIsDeletedFalse(Pageable pageable);
+    // 書籍一覧取得
+    Page<Book> findByIsDeletedFalse(Pageable pageable);
 
-       // タイトル検索
-       Page<Book> findByTitleContainingAndIsDeletedFalse(String keyword, Pageable pageable);
+    // タイトル検索
+    Page<Book> findByTitleContainingAndIsDeletedFalse(String keyword, Pageable pageable);
 
-       // 指定されたジャンルIDのリストを取得（OR条件）
-       Page<Book> findDistinctByGenres_IdInAndIsDeletedFalse(List<Long> genreIds,
-                     Pageable pageable);
+    // 指定されたジャンルIDのリストを取得（OR条件）
+    Page<Book> findDistinctByGenres_IdInAndIsDeletedFalse(List<Long> genreIds, Pageable pageable);
 
-       // 指定されたジャンルIDのリストを取得（AND条件）
-       @Query("""
-                     SELECT b FROM Book b
-                     JOIN b.genres bg
-                     WHERE bg.id IN :genreIds
-                     GROUP BY b.id
-                     HAVING COUNT(DISTINCT bg.id) = :size
-                     """)
-       Page<Book> findBooksHavingAllGenres(@Param("genreIds") List<Long> genreIds,
-                     @Param("size") long size, Pageable pageable);
+    // 指定されたジャンルIDのリストを取得（AND条件）
+    @Query("""
+            SELECT b FROM Book b
+            JOIN b.genres bg
+            WHERE bg.id IN :genreIds
+            GROUP BY b.id
+            HAVING COUNT(DISTINCT bg.id) = :size
+            """)
+    Page<Book> findBooksHavingAllGenres(@Param("genreIds") List<Long> genreIds,
+            @Param("size") Integer size, Pageable pageable);
 }
